Transformation of visually estimated cover (%) in a phytosociological survey to cover-abundance values

Species cover visually estimated in a phytosociological survey are transformed to:

Braun-Blanquet cover-abundance scale (Braun-Blanquet,1932)

Van Der Maarel cover-abundance scale (Van Der Maarel, 1978)

Dominance percentage (Tasser and Tappeiner, 2005)

see @details

Usage

PhytoCover(database, method)

Arguments

database

database with visually estimated cover of plant species. Rare species are added as '+". Rows are species and columns are surveys. The first column of the database reports the species names (Important: no space among words must exist). Database class must be data.frame

method

"BrBl": transform cover visually estimated to the Braun-Blanquet cover-abundance scale (see @details )

"Maarel": transform cover visually estimated to the Van Der Maarel cover-abundance scale (see @details )

"TassTapp": transform cover visually estimated to dominance percentage according to Tasser and Tappeiner (2005) (see @details )

Value

database with cover-abundance values. Rows are species and columns are surveys

Details

Conversion table

Visually estimated cover (%)	Braun Blanquet	Van Der Maarel	Tasser and Tappeiner

0	0	0	0

<1	+	2	0.3

1 - 5	1	3	2.8

6 -15	2a	5	10

16-25	2b	5	20.5

26-50	3	7	38

51-75	4	8	63

76-100	5	9	88

References

Braun-Blanquet J (1932) Plant sociology. The study of plant communi- ties. GD Fuller and HS Conard (Eds.). Authorized English translations of 'Pflanzensoziologie'. 1st ed. Printed in the United States of America. New York and London: McGraw-Hill Book Co. Inc.

Van der Maarel, E. (1979). Transformation of cover-abundance values in phytosociology and its effects on community similarity. Vegetatio, 39(2), 97-114.

Tasser E, Tappeiner U (2005) New model to predict rooting in diverse plant community compositions. Ecological Modelling 185:195-211.

Examples